Refactoring to a shared helper
