+++ ingestion_tools/scripts/data_validation/source/test/test_gain.py
@@ -1,5 +1,3 @@
-import warnings
-
 import pytest
 import tifffile
 from importers.gain import GainImporter
@@ -7,8 +5,7 @@ from importers.run import RunImporter
 from mrcfile.mrcinterpreter import MrcInterpreter
 
 from common.fs import S3Filesystem
-from data_validation.shared.helper.helper_mrc import HelperTestMRCHeader
-from data_validation.shared.helper.helper_tiff_mrc import helper_tiff_mrc_consistent
+from data_validation.shared.helper.frame_helper import GainTestHelper
 from data_validation.shared.util import (
     MRC_EXTENSION,
     PERMITTED_GAIN_EXTENSIONS,
@@ -20,7 +17,7 @@ PERMITTED_SOURCE_GAIN_EXTENSIONS = PERMITTED_GAIN_EXTENSIONS + [".dm4"]
 
 @pytest.mark.gain
 @pytest.mark.parametrize("run", pytest.cryoet.runs, ids=[ts.name for ts in pytest.cryoet.runs], scope="session")
-class TestGain(HelperTestMRCHeader):
+class TestGain(GainTestHelper):
 
     @pytest.fixture(scope="class")
     def gain_files(self, run: RunImporter) -> list[str]:
@@ -40,39 +37,4 @@ class TestGain(HelperTestMRCHeader):
     def set_helper_test_mrc_header_class_variables(
             self, gain_headers: dict[str, list[tifffile.TiffPage]| MrcInterpreter],
     ):
-        self.spacegroup = 0  # 2D image
         self.mrc_headers = gain_headers
-
-    ### DON'T RUN SOME MRC HEADER TESTS ###
-    def test_nlabel(self):
-        pytest.skip("Not applicable for gain files")
-
-    def test_nversion(self):
-        pytest.skip("Not applicable for gain files")
-
-    def test_mrc_spacing(self):
-        pytest.skip("Not applicable for gain files")
-
-    ### BEGIN Self-consistency tests ###
-    def test_extensions(self, gain_files: list[str]):
-        errors = []
-
-        for gain_file in gain_files:
-            if not any(gain_file.endswith(ext) for ext in PERMITTED_GAIN_EXTENSIONS):
-                errors.append(f"Invalid gain file extension: {gain_file}")
-
-        if errors:
-            warnings.warn("\n".join(errors), stacklevel=2)
-
-
-    def test_consistent(self, gain_headers: dict[str, list[tifffile.TiffPage]| MrcInterpreter]):
-        return helper_tiff_mrc_consistent(gain_headers)
-
-    def test_gain_nz(self):
-        def check_nz(header, _interpreter, _mrc_filename):
-            del _interpreter, _mrc_filename
-            assert header.nz == 1  # 2D image
-
-        self.mrc_header_helper(check_nz)
-
-    ### END Self-consistency tests ###
